Apply post-processing rules to formatting.
Country models support two operators to remove tokens from the hierarchy,
`cut-off-tokens` and `cut-off-children`, which remove individual elements
of entire subtrees.

This CL makes sure that if either an individual element or an entire subtree are
removed from the model, also the corresponding nodes from the formatting rules
are removed.