Introduce temporary outputs to deal with non-unique capture groups.
If we want to capture alternative patterns (e.g. 'andar \d+' and '\d+ andar'),
we may end up with colliding capture groups (both would be 'floor'). This CL
introduces the concept of a temporary output that is used as a capture group
name. After evaluating an expression, the renaming is reverted.